Removed transactional configuration and updated integration tests

+++ src/main/java/uk/gov/companieshouse/charges/data/service/ChargesService.java
@@ -5,13 +5,11 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Predicate;
-
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 import uk.gov.companieshouse.api.charges.ChargeApi;
 import uk.gov.companieshouse.api.charges.ChargesApi;
@@ -61,7 +59,6 @@ public class ChargesService {
      * @param chargeId      charges Id.
      * @param requestBody   request body.
      */
-    @Transactional
     public void upsertCharges(String contextId, String companyNumber, String chargeId,
                               InternalChargeApi requestBody) {
         Optional<ChargesDocument> chargesDocumentOptional = chargesRepository.findById(chargeId);
@@ -156,11 +153,6 @@ public class ChargesService {
 
     private void saveAndInvokeChsKafkaApi(String contextId, String companyNumber,
                                           String chargeId, ChargesDocument charges) {
-        this.chargesRepository.save(charges);
-        logger.debug(
-                String.format("Invoking chs-kafka-api for chargeId %s company number %s ",
-                        chargeId,
-                        companyNumber));
         ApiResponse<Void> res = chargesApiService.invokeChsKafkaApi(contextId,
                 companyNumber,
                 chargeId);
@@ -169,6 +161,11 @@ public class ChargesService {
             throw new ResponseStatusException(httpStatus != null
                     ? httpStatus : HttpStatus.INTERNAL_SERVER_ERROR, "invokeChsKafkaApi");
         }
+        this.chargesRepository.save(charges);
+        logger.debug(
+                String.format("Invoking chs-kafka-api for chargeId %s company number %s ",
+                        chargeId,
+                        companyNumber));
     }
 
     /**
@@ -205,13 +202,20 @@ public class ChargesService {
                                     HttpStatus.NOT_FOUND,
                                     "ChargeApi object doesn't exist for" + chargeId));
 
-            chargesApiService.invokeChsKafkaApiWithDeleteEvent(contextId,
+            ApiResponse<Void> apiResponse = chargesApiService.invokeChsKafkaApiWithDeleteEvent(
+                    contextId,
                     chargeId,
                     companyNumber,
                     chargeApi);
 
             logger.info(String.format("ChsKafka api invoked successfully for "
                     + "charge id %s and x-request-id %s", chargeId, contextId));
+
+            if (apiResponse == null || apiResponse.getStatusCode() != 200) {
+                throw new ResponseStatusException(apiResponse != null
+                        ? HttpStatus.valueOf(apiResponse.getStatusCode()) :
+                        HttpStatus.INTERNAL_SERVER_ERROR, "invokeChsKafkaApi");
+            }
 
             chargesRepository.deleteById(chargeId);
             logger.info(String.format(
